### 2020-02-09 #77 @whedon
[PRE REVIEW]: The Gene School: Metagenomics
**Submitting author:** @koadman (<a href="http://orcid.org/0000-0003-2397-7925">Aaron Darling</a>)
**Repository:** <a href="https://github.com/thegeneschool/metagenomics" target ="_blank">https://github.com/thegeneschool/metagenomics</a>
**Version:** v0.0.1

@koadman if you have any suggestions for potential reviewers then please mention them here in this thread (without tagging them with an @). In addition, [this list of people](https://bit.ly/jose-reviewers) have already agreed to review for JOSE and may be suitable for this submission.
