### 2020-02-25 #77 @emckiernan
Hi, @AstrobioMike, you signed up to review for JOSE. Would you like to contribute a review for this submission? The title is "The Gene School: Metagenomics". Let me know if you're available and willing. Thanks!
