### 2020-02-26 #77 @hughshanahan
Hi Erin
            yes please. To note I won’t be able to get to this until next week.

Hugh
__________________________
Hugh Shanahan
Professor of Open Science
( he | his )
Hugh.Shanahan@rhul.ac.uk<mailto:Hugh.Shanahan@rhul.ac.uk>
http://www.shanahanlab.org
@hughshanahan
Skype hugh_shanahan
Tel +44 (0)1784 443433
orcid.org/0000-0003-1374-6015<http://orcid.org/0000-0003-1374-6015>

Emails sent outside of office hours are not an expectation that you should process this outside of office hours.
