### 2020-02-28 #79 @whedon
[REVIEW]: The Gene School: Metagenomics
**Submitting author:** @koadman (<a href="http://orcid.org/0000-0003-2397-7925">Aaron Darling</a>)
**Repository:** <a href="https://github.com/thegeneschool/metagenomics" target ="_blank">https://github.com/thegeneschool/metagenomics</a>
**Version:** v0.0.1
