### 2020-02-28 #79 @emckiernan
Hi @hughshanahan, @pschloss: Thank you for agreeing to review for JOSE! This is where the action happens: please work your way through the review checklist, feel free to ask questions or post comments here, and also open issues in the submission repository as needed. I'm here to help, so please let me know if you need anything. Thanks!
