### 2020-03-27 #79 @pschloss
Can @emckiernan confirm that this is the correct review template? It seems to be for software rather than for teaching materials. I was expecting something more like what is given at #78
