### 2020-03-27 #79 @pschloss
Also... what is the preferred way of leaving feedback on the authors' repository? One issue with checkboxes or separate issues for each point?
