### 2020-03-30 #79 @emckiernan
> Can @emckiernan confirm that this is the correct review template? It seems to be for software rather than for teaching materials. I was expecting something more like what is given at #78

Hi @pschloss, thanks for your question. I see your point, the review checklist is different from that provided at #78, and is missing sections like the one on pedagogy. This was automatically generated by whedon, so I'm not sure why it is different. Since this is my first time editing for JOSE, I'd like to loop in @labarba, just to make sure we're on the right track. Lorena, do you know why we have a different review template here? Thanks for any help!
