### 2020-03-30 #79 @emckiernan
> Also... what is the preferred way of leaving feedback on the authors' repository? One issue with checkboxes or separate issues for each point?

Hi @pschloss: The guidelines say, "Comments in the REVIEW issue should be kept brief, as much as possible, with more lengthy suggestions or requests posted as separate issues, directly in the submission repository." So I'd say, if the comments are short, you can leave something like a checklist, but if the comments are more lengthy, I'd open these as separate issues. You can link back to all of these in your review. Hope that helps!
