### 2020-03-30 #79 @labarba
The checklist is auto-generated on the basis of the submission type chosen by the authors on the submission web form.

@koadman — do you remember choosing the submission type from the drop-down menu back when you submitted?
