### 2020-03-30 #79 @labarba
> what is the preferred way of leaving feedback on the authors' repository? One issue with checkboxes or separate issues for each point?

You can decide how you want to structure your feedback in the submission repository (whether to open one or several issues). But be sure to post here a link to the issue with a brief note, to create a cross-link.

Note that this Review thread will be archived together with the paper at publication time, for posterity.
