### 2020-04-01 thegeneschool/metagenomics#3 @pschloss
Peer review for JOSE
This manuscript describes a set of instructional materials for teaching researchers how to process shotgun metagenomic sequence data using traditional and Hi-C approaches. I haven't completed the checklist [over at JOSE](https://github.com/openjournals/jose-reviews/issues/79), because I think the one the authors selected was incorrect. I may have additional comments once I see the correct checklist.

I have a number of comments for the developers. My primary critique is that the materials are too personal to the developers use. There are several places where installation/setup details are incomplete, background material is omitted. I think closing these gaps will be important to ensuring the usability of the materials more broadly. Details of this are included in the following comments:

- [ ] Some introduction to the material is needed. The tutorial jumps right into logging into Jupyter Hub. What is metagenomics? What is the goal of this tutorial? What are the questions that this pipeline will allow us to answer? It seems that this tutorial is taking an assembly rather than a single read approach to analysis. I think this needs to be made clear to a learner.

- [ ] The nav bar on the left says "Jupyter setup" - I think it means to say "Setting up a working environment". That is the title of the page that is linked to.

- [ ] The link in the setup page to the QC page points to https://thegeneschool.github.io/qc rather than to https://thegeneschool.github.io/metagenomics/qc/. It would be good to run [html-proofer](https://github.com/gjtorikian/html-proofer) on the site to make sure all of the links work.

- [ ] The Jupyter Setup page ends with a comment about how to get started if the learner does not have access to AWS. My sense of looking through the material was that the pipeline would be done locally and not on AWS. I think the developers either need to remove that closing section or add a link/instructions on how to set up an instance on AWS for this pipeline. If the intention is to use AWS, then why not set up an AMI that has many of the resources already installed to make life easier for the learner? Still later in the Visualization page, it mentions "we are working on VMs", but the tutorial never creates a VM.

- [x] #7

- [ ] #8

- [ ] In the "Taxonomic analysis with other tools" section, it would be nice to have a brief comment on why you chose Metaphlan2 rather than these other options. There is a nice justification later on for metabat2 that would be a good model for this.

- [ ] It would be helpful to have a page that describes the conventions used throughout the tutorial. A description of what/why things are in monospace, what the orange boxes with warning symbols represent, what the purple boxes represent, and what the teal boxes represent would help. I'm sure some of this is obvious to people, but it would help with clarity. Also, the use of monospace in the text is inconsistent. For example, in some places, program names are monospace (e.g. kraken) and in others they are in normal font (e.g. metaphlan2)

- [ ] I like the challenge questions on the QC and Visualization pages! Could the developers include other challenge questions on the other pages? Providing links to a set of answers would help learners that are following the tutorial autonomously.

- [ ] The developers are going to use megahit, but also mention metaspades. Similar to the choice of metaphlan2, I think the learners would appreciate a justification. There is a nice justification later on for metabat2 that would be a good model for this.

- [ ] There don't appear to be instructions on how to install Singularity to get the Docker images

- [ ] I like the justification for metabat2. But what if a learner doesn't have time-series data, is it still what would be recommended? I suspect a cross-section study design is more common and having that information and perhaps a parallel example of what commands to use would be useful.

- [ ] The inclusion of the Hi-C page seems a bit out of place relative to the scope and context of the other pages. For example, they go into detail for how a Hi-C library is generated and where to go to get it done - there isn't a parallel description for traditional metagenomic sequencing. As I mention above, I think this information is needed within the other pages, but this still feels a little out of place. Perhaps the developers could pitch this page more as an "advanced move"? Something like, "now that a learner has completed the pipeline for traditional metagenomic data, let's see how to apply those concepts to a different way of doing metagenomics". Alternatively, the developers might consider making a separate Hi-C tutorial.

- [ ] Another incongruity for the Hi-C page relative to the others is the inclusion of references. This has me wondering why not provide references on the other pages where they make recommendations. The references here are also not uniformly formatted - there are two #1s and the first is flush left while the others have a hanging indent.

- [ ] I sympathize with the developers not wanting to reinvent the anvio tutorials, but the visualization page needs more detail on what they mean by a bin with errors and the steps of how to fix these. For learners, doing these steps with the data they are using in this tutorial would be beneficial. I feel like a page with visualization in the title should at least have one visualization :)

- [ ] There doesn't appear to be a functioning index page at https://thegeneschool.github.io - it gives a 404 error.

- [ ] The submitted version number (v0.1) is not the same as that on the repository (v1.0)

- [ ] The manuscript indicates that one of the learning objectives is to "Learn how a pilot study can inform experimental design". This is not covered in the online materials in meaningful depth.

- [ ] The manuscript points out that when this was previously deployed many of the intermediate files were pre-computed. For people doing this on their own or setting up the workshop to be taught locally, it would be helpful to include instructions for how to set up the pre-computed materials for learners. Furthermore, the learners need an indication in the tutorial of how long the more computationally intensive steps will take.

- [ ] If the goal of these materials is for an expert to be able to present the materials to their own learners, more details are needed alongside the tutorial to help the instructor. Set up details, time to completion, common hang-up points, etc would significantly improve the usability of the materials.

- [ ] The PDF version of the manuscript did not correctly render the reference for "Visualization of bins and bin refinement with anvi'o" in the Learning Objectives.

- [ ] The README could provide a link and reference to the metagenomic time-series and Hi-C dataset.
