### 2020-04-01 #79 @pschloss
I have posted my review in the [authors' repository](https://github.com/thegeneschool/metagenomics/issues/3). I have not completed the checklist above because I'm pretty confident it's the wrong one. I may have additional comments if a different checklist is given.
