### 2020-04-02 #79 @kyleniemeyer
Oh, good point. It looks like whedon posted the JOSS review checklist, rather than the JOSE checklist. @arfon @labarba ?
