### 2020-04-03 #79 @arfon
> Oh, good point. It looks like whedon posted the JOSS review checklist, rather than the JOSE checklist. @arfon @labarba ?

Checking...
