### 2020-04-03 #79 @arfon
@kyleniemeyer @labarba - looks like this is marked as `software` in the JOSE database so I think the review checklist is correct. Were you expecting the checklist for a `learning module`?
