### 2020-04-03 #79 @kyleniemeyer
@arfon yeah, I think that was a mistake that we didn't catch. This looks to be a learning module instead. Is that an easy fix, or do we just need to manually change the lists?
