### 2020-04-03 #79 @arfon
I'm afraid this is a manual change to the lists so I think you are as well-positioned as I am to make this change :-)
