### 2020-04-03 #79 @pschloss
I've completed the checklist. I think it matches what I have written in my review in the repository.
