### 2020-05-13 #79 @emckiernan
Hi @erahulkulkarni! You signed up to review for JOSE. Would you like to contribute a review for this submission? The title is "The Gene School: Metagenomics". Please let me know if you're available and willing. Thanks!
