### 2020-05-14 #79 @erahulkulkarni
Good Morning  Erin C. McKiernan,

I am a bit rusty on BioInformatics since it has been a long time I have
worked on any projects.

So, I may not be the best match at present.

If it is still okay, please let me know.

Thanks and Regards,
*Rahul Kulkarni*

On Wed, May 13, 2020 at 11:06 PM Erin C. McKiernan <notifications@github.com>
wrote:

> Hi @erahulkulkarni <https://github.com/erahulkulkarni>! You signed up to
> review for JOSE. Would you like to contribute a review for this submission?
> The title is "The Gene School: Metagenomics". Please let me know if you're
> available and willing. Thanks!
>
