### 2020-05-19 #79 @emckiernan
Hi @erahulkulkarni, thanks for your response and apologies for my slow reply! If you're still willing, it would be great to have you review, even if rusty :). I think one of the most important things is to evaluate that all the documentation of the software and educational tools described is complete and makes sense to you. In my opinion, the didactic materials should be as clear as possible, even to someone without in-depth expertise in the area. Hope that helps, and thanks in advance for considering!
