### 2020-05-20 #79 @erahulkulkarni
Good Morning Erin C. McKiernan,

No worries.

Okay then, I will check rules / documentation of review and a give a try.
Thank You.

Thanks and Regards,
*Rahul Kulkarni*

On Wed, May 20, 2020 at 12:16 AM Erin C. McKiernan <notifications@github.com>
wrote:

> Hi @erahulkulkarni <https://github.com/erahulkulkarni>, thanks for your
> response and apologies for my slow reply! If you're still willing, it would
> be great to have you review, even if rusty :). I think one of the most
> important things is to evaluate that all the documentation of the software
> and educational tools described is complete and makes sense to you. In my
> opinion, the didactic materials should be as clear as possible, even to
> someone without in-depth expertise in the area. Hope that helps, and thanks
> in advance for considering!
>
