### 2020-05-22 #79 @erahulkulkarni
Thank You.

Thanks and Regards,
*Rahul Kulkarni*

On Fri, May 22, 2020 at 1:43 PM Lorena A. Barba <notifications@github.com>
wrote:

> @emckiernan <https://github.com/emckiernan> -- @whedon is not able to
> edit the lead comment in this thread to add a new review checklist, so this
> step is manual. I just edited the comment and added a checklist for
> @erahulkulkarni <https://github.com/erahulkulkarni> (thanks for agreeing
> to review!)
>
