### 2020-06-02 #79 @erahulkulkarni
Good Morning,

Sorry due to academic uncertainty, I am unable to concentrate. I am not
sure if I can give my best in review.

Thanks and Regards,
*Rahul Kulkarni*

On Fri, May 22, 2020 at 3:49 PM Rahul Kulkarni <erahulkulkarni@gmail.com>
wrote:

> Thank You.
>
>
> Thanks and Regards,
> *Rahul Kulkarni*
>
> On Fri, May 22, 2020 at 1:43 PM Lorena A. Barba <notifications@github.com>
> wrote:
>
>> @emckiernan <https://github.com/emckiernan> -- @whedon is not able to
>> edit the lead comment in this thread to add a new review checklist, so this
>> step is manual. I just edited the comment and added a checklist for
>> @erahulkulkarni <https://github.com/erahulkulkarni> (thanks for agreeing
>> to review!)
>>
>> —
>> You are receiving this because you were mentioned.
>> Reply to this email directly, view it on GitHub
>> <https://github.com/openjournals/jose-reviews/issues/79#issuecomment-632561588>,
>> or unsubscribe
>> <https://github.com/notifications/unsubscribe-auth/ABBG2GKMUA6WXD7JOWJTCZ3RSYX27ANCNFSM4K53OZEA>
>> .
>>
>
