### 2020-06-09 #79 @erahulkulkarni
Thank you.

Sorry for the trouble.

Thanks and Regards,
*Rahul Kulkarni*

On Mon, Jun 8, 2020 at 11:02 PM Erin C. McKiernan <notifications@github.com>
wrote:

> @whedon <https://github.com/whedon> remove @erahulkulkarni
> <https://github.com/erahulkulkarni> as reviewer
>
