### 2020-08-16 #79 @emckiernan
> Hi @emckiernan - yes I can review this. The thread here is quite long though - can you please provide clarity on which template to complete and procedure to use?

Great, thanks so much, @pvanheus! I will add you as a reviewer and then manually add a checklist for you above. If you still have questions, just let me know and I'll be happy to help. Thanks again!
