### 2020-08-31 #79 @emckiernan
Hi @pvanheus, just checking in! Was the checklist clear? Is there anything I can help with?
