### 2020-09-23 #79 @fbidu
> Hi @fbidu! You signed up to review for JOSE. Would you like to contribute a review for this submission? The title is "The Gene School: Metagenomics". Please let me know if you're available and willing. Thanks!

Yes, I am! I think that I have a friend that may be interested as well, can I invite him somehow?

Thanks
