### 2020-09-23 #79 @emckiernan
> Yes, I am! I think that I have a friend that may be interested as well, can I invite him somehow?
>
> Thanks

Great, thanks so much @fbidu! I'll add you as a reviewer and then insert a checklist for you above. Could you please send the contact information for your friend to my email emck31 at gmail? I can look it over and invite him if we need an extra reviewer. Thanks again!
