### 2020-09-28 #79 @pvanheus
Hi @emckiernan - apologies, I have been ill. I think there is a problem with the template you created though - the URL link points to https://github.com/Nikoleta-v3/Game-Theory-and-Python - which is not the repo being discussed here - at the top of the page https://github.com/thegeneschool/metagenomics is mentioned. What is going on?
