### 2020-09-28 #79 @emckiernan
> Hi @emckiernan - apologies, I have been ill. I think there is a problem with the template you created though - the URL link points to https://github.com/Nikoleta-v3/Game-Theory-and-Python - which is not the repo being discussed here - at the top of the page https://github.com/thegeneschool/metagenomics is mentioned. What is going on?

Hi @pvanheus, no worries at all, I'm sorry to hear you were ill. I hope you're feeling better! And thank you for flagging the URL error. I'm not sure why the checklist originally generated by the bot linked to that repo. After that it was copy-paste errors when checklists were manually added. But you're right, the correct repository for review is https://github.com/thegeneschool/metagenomics. I can correct the links above.
