### 2020-12-11 #79 @pvanheus
@emckiernan I have submitted a much delayed [review](https://github.com/thegeneschool/metagenomics/issues/6). I must admit that reviewing this material was not easy, but working through your checklist and considering the material from the perspective of someone who might want to adopt it helped in the end. I'll also note that I opened two issues ([here](https://github.com/thegeneschool/metagenomics/issues/4) and [here](https://github.com/thegeneschool/metagenomics/issues/5)) that have not been addressed by the authors since October.
