### 2020-12-21 #79 @koadman
Thanks all, yes ready to move forward with revisions in the new year. Plenty to do here, thanks for the helpful feedback!
